moved all unique landing components to Landing
