+++ src/client/pages/Landing/UpComingEventCard.js
